完成用户管理，以及部分bug修复等

+++ oauth2-server-boot/src/main/java/com/newnil/cas/oauth2/provider/controller/RoleAdminController.java
@@ -1,6 +1,7 @@
 package com.newnil.cas.oauth2.provider.controller;
 
 import com.newnil.cas.oauth2.provider.dao.entity.RoleEntity;
+import com.newnil.cas.oauth2.provider.dao.entity.UserRoleXRef;
 import com.newnil.cas.oauth2.provider.dao.repository.RoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -15,13 +16,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.regex.Pattern;
 
+import static com.newnil.cas.oauth2.provider.webhelper.RedirectMessageHelper.*;
+
 @Controller
 @RequestMapping("/roles")
-@PreAuthorize("hasRole('ROLE_USER')")
+@PreAuthorize("hasRole('ROLE_ADMIN')")
 public class RoleAdminController {
 
     @Autowired
@@ -34,25 +36,24 @@ public class RoleAdminController {
         return "roles/roles";
     }
 
-    private static final Pattern ROLE_NAME_PATTERN = Pattern.compile("^[A-Z_]+$");
+    private static final Pattern ROLE_NAME_PATTERN = Pattern.compile("^[a-zA-Z0-9_]+$");
 
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE,
             produces = {MediaType.TEXT_HTML_VALUE, MediaType.APPLICATION_XHTML_XML_VALUE})
     public String createRole(@RequestParam("roleName") String roleName, RedirectAttributes attributes) {
         if (ROLE_NAME_PATTERN.matcher(roleName).matches()) {
             if (roleRepository.findOneByName(roleName.toUpperCase()).isPresent()) {
                 // error message
-                attributes.addFlashAttribute("dangerMessages", Collections.singletonList("角色名 " + roleName + " 已存在。"));
+                addErrorMessage(attributes, "角色名 " + roleName + " 已存在。");
                 attributes.addFlashAttribute("roleName", roleName);
             } else {
                 roleRepository.save(RoleEntity.builder().name(roleName.toUpperCase()).build());
                 // success message
-                attributes.addFlashAttribute("successMessages", Collections.singletonList("已成功添加 " + roleName + " 角色。"));
+                addSuccessMessage(attributes, "已成功添加 " + roleName + " 角色。");
             }
 
         } else {
-            attributes.addFlashAttribute("dangerMessages", Collections.singletonList("角色名 " + roleName + " 含有非法字符。（只能使用[A-Z_]）"));
+            addErrorMessage(attributes, "角色名 " + roleName + " 含有非法字符。（只能使用[a-zA-Z0-9_]）");
             attributes.addFlashAttribute("roleName", roleName);
         }
         return "redirect:/roles";
@@ -61,24 +62,31 @@ public class RoleAdminController {
     private static final String[] INVINCIBLE_ROLES = {"ADMIN", "USER"};
     private static final List<String> INVINCIBLE_ROLES_LIST = Arrays.asList(INVINCIBLE_ROLES);
 
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @RequestMapping(path = "/_remove/{roleName}", method = RequestMethod.GET, produces = {MediaType.TEXT_HTML_VALUE, MediaType.APPLICATION_XHTML_XML_VALUE})
     public String removeRole(@PathVariable("roleName") String roleName, RedirectAttributes attributes) {
 
         if (INVINCIBLE_ROLES_LIST.contains(roleName.toUpperCase())) {
 
-            attributes.addFlashAttribute("dangerMessages", "该角色不可删除：" + roleName);
+            addErrorMessage(attributes, "该角色不可删除：" + roleName);
 
         } else {
 
             roleRepository.findOneByName(roleName.toUpperCase()).map(
                     roleEntity -> {
-                        roleRepository.delete(roleEntity);
-                        attributes.addFlashAttribute("successMessages", Collections.singletonList("已成功删除 " + roleName + " 角色。"));
+
+                        List<UserRoleXRef> xRefList = roleEntity.getUsers();
+                        if (xRefList.isEmpty()) {
+
+                            roleRepository.delete(roleEntity);
+                            addSuccessMessage(attributes, "已成功删除 " + roleName + " 角色。");
+                        } else {
+                            addErrorMessage(attributes, "有" + xRefList.size() + "个用户正在使用该角色，无法删除。");
+                        }
+
                         return roleEntity;
                     }
             ).orElseGet(() -> {
-                attributes.addFlashAttribute("warningMessages", Collections.singletonList("没有找到 " + roleName + " 角色。"));
+                addWarningMessage(attributes, "没有找到 " + roleName + " 角色。");
                 return null;
             });
         }
